refs #46081 - throw exception if processes exists with availability to delete

+++ src/Zmsapi/AvailabilityDelete.php
@@ -23,10 +23,13 @@ class AvailabilityDelete extends BaseController
         (new Helper\User($request))->checkRights();
         $query = new Query();
         $entity = $query->readEntity($args['id'], 2);
+        $processList = (new \BO\Zmsdb\Process())->readListFromAvailability($entity);
+        if ($processList->count()) {
+            throw new Exception\Availability\AvailabilityHasProcess();
+        }
+
         if ($entity->hasId() && $query->deleteEntity($entity->getId())) {
-            (new \BO\Zmsdb\Slot)->writeByAvailability($entity, \App::$now);
             (new \BO\Zmsdb\Helper\CalculateSlots(\App::DEBUG))->writePostProcessingByScope($entity->scope, \App::$now);
-            \BO\Zmsdb\Connection\Select::writeCommit();
         }
 
         $message = Response\Message::create($request);
